:recycle: Refactor producer application and adjust consumer listener
The producer application has been completely refactored with modifications encompassing the deletion of the old producer class, addition of a new service class, and adjustments in the corresponding controller. The listener in the consumer application has been slightly edited for better logger usage. Lastly, the application config file and the message model in the producer application have been updated to fit the new requirements.